requirements.txt
python-telegram-bot>=22.0
apscheduler>=3.10.4
libsql>=0.3.0
